+++ src/Schema/Definitions/FieldDefinition.php
@@ -56,7 +56,7 @@ class FieldDefinition {
 	private $autoIncrement;
 
 	const TYPE_TINYINT = 'tinyint';
-	const TYPE_TEXT = 'str'; // need at least short sting vs text vs blob
+	const TYPE_BLOB = 'blob'; // need at least short sting vs text vs blob
 	const TYPE_INTEGER = 'int';
 	const TYPE_DECIMAL = 'decimal';
 	const TYPE_BIGINT = 'bigint';