+++ tests/phpunit/SQLite/SQLiteFieldSqlBuilderTest.php
@@ -92,7 +92,7 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'fieldName',
-				FieldDefinition::TYPE_TEXT,
+				FieldDefinition::TYPE_BLOB,
 				FieldDefinition::NOT_NULL,
 				'foo'
 			),