+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@ is thus responsible for autoloading.
 * MediaWiki plugin: added compatibility with changes in MediaWiki 1.22.
 * Replaced ResultIterator by PHPs native Iterator interface. The interface signature is the same.
 * FieldDefinition::TYPE_BOOL has been changed to FieldDefinition::TYPE_TINYINT
+* FieldDefinition::TYPE_TEXT has been changed to FieldDefinition::TYPE_BLOB
 
 ### Additions